BROKEN upgrade to latest repo code

+++ src/Mlx90640/MLX90640_I2C_Driver.cpp
@@ -19,13 +19,15 @@
 #include <M5Unified.h>
 #include <Wire.h>
 
-void MLX90640_I2CInit() {
+void MLX90640_I2CInit()
+{
 }
 
 // Read a number of words from startAddress. Store into Data array.
 // Returns 0 if successful, -1 if error
 int MLX90640_I2CRead(uint8_t _deviceAddress, unsigned int startAddress,
-                     unsigned int nWordsRead, uint16_t *data) {
+                     unsigned int nWordsRead, uint16_t *data)
+{
     // Caller passes number of 'unsigned ints to read', increase this to 'bytes
     // to read'
     uint16_t bytesRemaining = nWordsRead * 2;
@@ -33,29 +35,32 @@ int MLX90640_I2CRead(uint8_t _deviceAddress, unsigned int startAddress,
     // It doesn't look like sequential read works. Do we need to re-issue the
     // address command each time?
 
-    uint16_t dataSpot = 0;  // Start at beginning of array
+    uint16_t dataSpot = 0; // Start at beginning of array
 
     // Setup a series of chunked I2C_BUFFER_LENGTH byte reads
-    while (bytesRemaining > 0) {
+    while (bytesRemaining > 0)
+    {
         Wire.beginTransmission(_deviceAddress);
-        Wire.write(startAddress >> 8);         // MSB
-        Wire.write(startAddress & 0xFF);       // LSB
-        if (Wire.endTransmission(false) != 0)  // Do not release bus
+        Wire.write(startAddress >> 8);        // MSB
+        Wire.write(startAddress & 0xFF);      // LSB
+        if (Wire.endTransmission(false) != 0) // Do not release bus
         {
             Serial.println("No ack read");
-            return (0);  // Sensor did not ACK
+            return (0); // Sensor did not ACK
         }
 
         uint16_t numberOfBytesToRead = bytesRemaining;
         if (numberOfBytesToRead > I2C_BUFFER_LENGTH)
             numberOfBytesToRead = I2C_BUFFER_LENGTH;
 
-        Wire.requestFrom((uint8_t)_deviceAddress, numberOfBytesToRead);
-        if (Wire.available()) {
-            for (uint16_t x = 0; x < numberOfBytesToRead / 2; x++) {
+        Wire.requestFrom((uint8_t)_deviceAddress, (uint8_t)numberOfBytesToRead);
+        if (Wire.available())
+        {
+            for (uint16_t x = 0; x < numberOfBytesToRead / 2; x++)
+            {
                 // Store data into array
-                data[dataSpot] = Wire.read() << 8;  // MSB
-                data[dataSpot] |= Wire.read();      // LSB
+                data[dataSpot] = Wire.read() << 8; // MSB
+                data[dataSpot] |= Wire.read();     // LSB
 
                 dataSpot++;
             }
@@ -66,25 +71,28 @@ int MLX90640_I2CRead(uint8_t _deviceAddress, unsigned int startAddress,
         startAddress += numberOfBytesToRead / 2;
     }
 
-    return (0);  // Success
+    return (0); // Success
 }
 
 // Set I2C Freq, in kHz
 // MLX90640_I2CFreqSet(1000) sets frequency to 1MHz
-void MLX90640_I2CFreqSet(int freq) {
+void MLX90640_I2CFreqSet(int freq)
+{
     // i2c.frequency(1000 * freq);
     Wire.setClock((long)1000 * freq);
 }
 
 // Write two bytes to a two byte address
 int MLX90640_I2CWrite(uint8_t _deviceAddress, unsigned int writeAddress,
-                      uint16_t data) {
+                      uint16_t data)
+{
     Wire.beginTransmission((uint8_t)_deviceAddress);
-    Wire.write(writeAddress >> 8);    // MSB
-    Wire.write(writeAddress & 0xFF);  // LSB
-    Wire.write(data >> 8);            // MSB
-    Wire.write(data & 0xFF);          // LSB
-    if (Wire.endTransmission() != 0) {
+    Wire.write(writeAddress >> 8);   // MSB
+    Wire.write(writeAddress & 0xFF); // LSB
+    Wire.write(data >> 8);           // MSB
+    Wire.write(data & 0xFF);         // LSB
+    if (Wire.endTransmission() != 0)
+    {
         // Sensor did not ACK
         Serial.println("Error: Sensor did not ack");
         return (-1);
@@ -92,10 +100,11 @@ int MLX90640_I2CWrite(uint8_t _deviceAddress, unsigned int writeAddress,
 
     uint16_t dataCheck;
     MLX90640_I2CRead(_deviceAddress, writeAddress, 1, &dataCheck);
-    if (dataCheck != data) {
+    if (dataCheck != data)
+    {
         // Serial.println("The write request didn't stick");
         return -2;
     }
 
-    return (0);  // Success
+    return (0); // Success
 }
